% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{add.branchSurvival}
\alias{add.branchSurvival}
\title{Function to add new column to Branches dataframe indicating for each branch whether or not it leaves any extant descendents}
\usage{
add.branchSurvival(Branches)
}
\arguments{
\item{Branches:}{tree in internal branch format}
}
\description{
Function to add new column to Branches dataframe indicating for each branch whether or not it leaves any extant descendents
}
\examples{
add.branchSurvival()
}
\keyword{branch,}
\keyword{extant}

