% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_branchSurvival}
\alias{add_branchSurvival}
\title{Function to add new column to Branches dataframe indicating for each branch whether or not it leaves any extant descendents}
\usage{
add_branchSurvival(Branches)
}
\arguments{
\item{Branches}{tree in internal branch format}
}
\description{
Function to add new column to Branches dataframe indicating for each branch whether or not it leaves any extant descendents
}
