% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{convert.2PbranchesToPhylo}
\alias{convert.2PbranchesToPhylo}
\title{Converting raw trees to phylo format}
\usage{
convert.2PbranchesToPhylo(P.PBranches, Q.PBranches, prune.extinct = FALSE)
}
\arguments{
\item{P.PBranches:}{Parasite-tree in raw matrix format}

\item{Q.PBranches:}{Parasite-tree in raw matrix format}

\item{prune.extinct:}{whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts raw dual parasite tree matricies into phylo format
}
\examples{
convert.2PbranchesToPhylo()
}
\keyword{convert,}
\keyword{format,}
\keyword{phylo}
