% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{convert.2PphyloToBranches}
\alias{convert.2PphyloToBranches}
\title{Convert cophylogenetic trees from Ape's phylo format to the internal Branches format}
\usage{
convert.2PphyloToBranches(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and two parasite trees}
}
\description{
The following function converts a phylo host-(dual) parasite tree into internal Branches format
}
\examples{
convert.2PphyloToBranches()
}
\keyword{Branches}
\keyword{convert,}
\keyword{format}
\keyword{format,}
