% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{convert.HbranchesToPhylo}
\alias{convert.HbranchesToPhylo}
\title{Converting raw tree to phylo format}
\usage{
convert.HbranchesToPhylo(HBranches, prune.extinct = FALSE)
}
\arguments{
\item{HBranches:}{Host-tree in raw matrix format}

\item{prune.extinct:}{whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts a raw host tree matrix into phylo format
}
\examples{
convert.HbranchesToPhylo()
}
\keyword{convert,}
\keyword{format,}
\keyword{phylo}

