% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{convert.PbranchesToPhylo}
\alias{convert.PbranchesToPhylo}
\title{Converting raw Parasite tree to phylo format}
\usage{
convert.PbranchesToPhylo(PBranches, prune.extinct = FALSE)
}
\arguments{
\item{PBranches:}{Parasite-tree in raw matrix format, prune.extinct: whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts a raw parasite tree matrix into phylo format
}
\examples{
convert.PbranchesToPhylo()
}
\keyword{convert,}
\keyword{format,}
\keyword{phylo}

