% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{convert.phyloToBranches}
\alias{convert.phyloToBranches}
\title{Convert cophylogenetic trees from Ape's phylo format to the internal Branches format}
\usage{
convert.phyloToBranches(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite tree}
}
\description{
The following function converts a phylo host-parasite tree into internal Branches format
}
\examples{
convert.phyloToBranches()
}
\keyword{Branches}
\keyword{convert,}
\keyword{format}
\keyword{format,}

