% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_HBtoPhy}
\alias{convert_HBtoPhy}
\title{Converting raw tree to phylo format}
\usage{
convert_HBtoPhy(HBranches, prune.extinct = FALSE)
}
\arguments{
\item{HBranches}{Host-tree in raw matrix format}

\item{prune.extinct}{whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts a raw host tree matrix into phylo format
}
