% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_HPBranchesToCophylo}
\alias{convert_HPBranchesToCophylo}
\title{Converting raw trees to phylo format}
\usage{
convert_HPBranchesToCophylo(HBranches, PBranches, prune.extinct = FALSE)
}
\arguments{
\item{HBranches:}{Host-tree in raw matrix format}

\item{PBranches:}{Parasite-tree in raw matrix format}

\item{prune.extinct:}{whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts raw host-parasite tree matricies into phylo format
}
\examples{
convert_HPBranchesToCophylo()
}
