% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_PBranchesToPhylo}
\alias{convert_PBranchesToPhylo}
\title{Converting raw Parasite tree to phylo format}
\usage{
convert_PBranchesToPhylo(PBranches, prune.extinct = FALSE)
}
\arguments{
\item{PBranches:}{Parasite-tree in raw matrix format}

\item{prune.extinct:}{whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts a raw parasite tree matrix into phylo format
}
\examples{
convert_PBranchesToPhylo()
}
\keyword{convert,}
\keyword{format,}
\keyword{phylo}
