% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_PQBranchesToPhylo}
\alias{convert_PQBranchesToPhylo}
\title{Converting raw trees to phylo format}
\usage{
convert_PQBranchesToPhylo(P.PBranches, Q.PBranches, prune.extinct = FALSE)
}
\arguments{
\item{P.PBranches}{Parasite-tree in raw matrix format}

\item{Q.PBranches}{Parasite-tree in raw matrix format}

\item{prune.extinct}{whether to remove all extinct branches (defaulting to FALSE)}
}
\description{
The following function converts raw dual parasite tree dataframes into phylo format
}
\examples{
Htree<-rphylo_H(tmax=5, export.format="raw")
HPQtree<-rcophylo_PQonH(H.tree=Htree, tmax=5, export.format="raw")
convert_PQBranchesToPhylo(P.PBranches=HPQtree[[2]],Q.PBranches=HPQtree[[2]])
}
