% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{cophy.2PonH.infectionResponse}
\alias{cophy.2PonH.infectionResponse}
\title{A random dual parasite tree building function with discrete responsive parasite resistance evolution}
\usage{
cophy.2PonH.infectionResponse(tmax, H.tree, beta = 0.1, gamma.P = 0.2,
  gamma.Q = 0.2, sigma.self = 0, sigma.cross = 0, mu.P = 0.5,
  mu.Q = 0.5, epsilon.1to0, epsilon.0to1, omega, rho, psi,
  TraitTracking = NA, P.startT = 0, ini.Hbranch = NA, Gdist = NA,
  prune.extinct = FALSE, export.format = "Phylo", timestep = 0.001,
  DBINC = 100)
}
\arguments{
\item{tmax:}{maximum time for which to simulate}

\item{H.tree:}{a pre-built host phylogenetic tree}

\item{beta:}{parasite host jump rate}

\item{gamma.P:}{dependency on genetic distance for host jumps}

\item{gamma.Q:}{dependency on genetic distance for host jumps}

\item{sigma.self:}{probability of successful co-infection with related parasite following host jump}

\item{sigma.cross:}{probability of successful co-infection with unrelated parasite following host jump}

\item{mu.P:}{parasite extinction rate}

\item{mu.Q:}{parasite extinction rate}

\item{prune.extinct:}{whether to remove all extinct branches defaulting to FALSE}

\item{export.format:}{either "Phylo" (exported in Ape Phylo format, the default setting)) or "Raw" (just a list of branches as used within the function itself)}

\item{P.startT:}{the timepoint at which a parasite invades the host-tree}

\item{ini.Hbranch:}{the host branch from which the parasite invasion is initiated (defaults to NA)}

\item{Gdist:}{can input a pre-calculated distance matrix of the living host branches at time of infection (defaults to NA), timestep: timestep for simulations}
}
\description{
The following function simulates a two coevolving parasite phylogenetic trees on a pre-built host phylogeny.
}
\examples{
cophy.2PonH.infectionResponse()
}
\keyword{Multi-Parasite}
\keyword{phylogeny}
