% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{cophylogeny}
\alias{cophylogeny}
\title{Creates a cophylogeny object}
\usage{
cophylogeny(HP.tree)
}
\arguments{
\item{HP.tree}{a list of a pre-built host phylogenetic tree and a parasite
phylogenetic tree of class 'cophylogeny' or 'data.frame'}
}
\value{
this function returns an object of class 'cophylogeny' which can be
  for plotting and printing pur.
}
\description{
This function creates an object of class 'cophylogeny', which can be passed
to \code{\link{plot.cophylogeny}}. This object must contain at least one
host and one parasite tree.
}
\examples{
HTree<-rphylo_H(tmax=5, exportFormat='raw')
HPTree<-rcophylo(HTree=HTree)
cophylogeny(HPTree)
}
\keyword{class}
\keyword{cophylogeny,}
