% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimal_places}
\alias{decimal_places}
\title{A function to count the number of decimal places in a number}
\usage{
decimal_places(x)
}
\arguments{
\item{x}{some number}
}
\description{
Taken from: https://stackoverflow.com/questions/5173692/how-to-return-number-of-decimal-places-in-r
}
\author{
daroczig
examples
decimal_places(1)
decimal_places(0.4)
}
