% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{get.2PEventsThroughTime}
\alias{get.2PEventsThroughTime}
\title{The following function returns a matrix containing counts of how often, in a given time step, different events have taken place within a parasite tree.}
\usage{
get.2PEventsThroughTime(cophy, tmin = 0, tmax = "max", dt = 1)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and two parasite trees}
}
\description{
The following function records parasite events from differnts lineages through time including: Start of time interval, End of time interval, Number of living branches (at the end of time interval), cospeciation events, host shifts, extinction (with hosts surviving), co-extinction (extinction caused by host extinction), speciation events for lineages with surviving descendents
}
\examples{
get.2PEventsThroughTime()
}
\keyword{cophylogeny,}
\keyword{events}
