% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.2Pinfectionlevels}
\alias{get.2Pinfectionlevels}
\title{The following function counts infection levels for two parasite lineages}
\usage{
get.2Pinfectionlevels(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and two parasite trees}
}
\description{
The following function counts the number of hosts infected with a particular number of parasites from the same parasite phylogeny for each parasite tree
}
\examples{
get.2Pinfectionlevels()
}
\keyword{cophylogeny,}
\keyword{infection-counts}

