% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.2Pinfectionstats}
\alias{get.2Pinfectionstats}
\title{The following function calculated basic statistics in dual parasite simulations}
\usage{
get.2Pinfectionstats(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and two parasite trees}
}
\description{
The following function calculates the basic statistics concerning a particular simulation (involving two parasites) including: number of extant host species, number of extant parasite species, percentage of host species that are infected by at least one parasite, mean number of parasites per host
}
\examples{
get.2Pinfectionstats()
}
\keyword{cophylogeny,}
\keyword{statistics}

