% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{get.branchesThroughTime}
\alias{get.branchesThroughTime}
\title{The following functioin returns a vector of branch (species) numbers through time, using timesteps dt and a maximum duration tmax that may be greater than the actual length of the tree.}
\usage{
get.branchesThroughTime(phy, tmax, dt)
}
\arguments{
\item{phy:}{a phylogeny (in phylo format) containing one tree}

\item{tmax:}{maximum time for which to simulate}

\item{dt:}{step size for performing calculations}
}
\description{
The following functioin returns a vector of branch (species) numbers through time, using timesteps dt and a maximum duration tmax that may be greater than the actual length of the tree.
}
\examples{
get.branchesThroughTime()
}
\keyword{branches}
\keyword{cophylogeny,}
