% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.Gdist}
\alias{get.Gdist}
\title{Calculating the distance matrix between host branches alive at a particular timepoint}
\usage{
get.Gdist(branches, t = NA)
}
\arguments{
\item{branches:}{raw branches matrix of a host tree}

\item{t:}{timepoint in simulation at which you want distance information}
}
\description{
Calculating the distance matrix between host branches alive at a particular timepoint
}
\examples{
get.Gdist()
}
\keyword{distance}
\keyword{genetic}

