% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.HostJumps}
\alias{get.HostJumps}
\title{The following function counts host-jumps}
\usage{
get.HostJumps(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite trees}
}
\description{
The following function counts the host-humps of a host-parasite phylogeny
}
\examples{
get.HostJumps()
}
\keyword{cophylogeny,}
\keyword{host-jumps}

