% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.infectionlevels}
\alias{get.infectionlevels}
\title{The following function counts infection levels}
\usage{
get.infectionlevels(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite trees}
}
\description{
The following function counts the number of hosts infected with a particular number of parasites from the same parasite phylogeny
}
\examples{
get.infectionlevels()
}
\keyword{cophylogeny,}
\keyword{infection-counts}

