% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.infectionstats}
\alias{get.infectionstats}
\title{The following function calculated basic statistics}
\usage{
get.infectionstats(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite trees}
}
\description{
The following function calculates the basic statistics concerning a particular simulation including: number of extant host species, number of extant parasite species, percentage of host species that are infected by at least one parasite, mean number of parasites per host
}
\examples{
get.infectionstats()
}
\keyword{cophylogeny,}
\keyword{statistics}

