% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.PEventsThroughTime}
\alias{get.PEventsThroughTime}
\title{The following function returns a matrix containing counts of how often, in a given time step, different events have taken place within a parasite tree.}
\usage{
get.PEventsThroughTime(cophy, tmin = 0, tmax = "max", dt = 1)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite trees}

\item{tmin:}{the timepoint in the simulation from which parasite events should be recorded, default = 0 (start point for the cophylogeny)}

\item{tmax:the}{timepoint in the simulation until which parasite events should be recorded, default = 'max' (end point for the cophylogeny)}

\item{dt:}{step size for the time points at which calculations are made}
}
\description{
The following function records parasites events through time including: Start of time interval, End of time interval, Number of living branches (at the end of time interval), cospeciation events, host shifts, extinction (with hosts surviving), co-extinction (extinction caused by host extinction), speciation events for lineages with surviving descendents
}
\examples{
get.PEventsThroughTime()
}
\keyword{cophylogeny,}
\keyword{events}

