% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.PextinctionTime}
\alias{get.PextinctionTime}
\title{Function to calculate the time of the last surviving parasite species}
\usage{
get.PextinctionTime(phy)
}
\arguments{
\item{phy:}{tree in phylo format}
}
\description{
Function to calculate the time of the last surviving parasite species
}
\examples{
get.PextinctionTime()
}
\keyword{last}
\keyword{parasite}

