% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.PHdist}
\alias{get.PHdist}
\title{Calculating the distance matrix between living parasite branches, as well that of associated hosts}
\usage{
get.PHdist(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite tree}
}
\description{
The following function returns a matrix of the partristic distances between all extant parasite species, and another matrix with the patristic distances of the associated host species.
}
\examples{
get.PHdist()
}
\keyword{distance}
\keyword{genetic}

