% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{get.PHdistCorrelation}
\alias{get.PHdistCorrelation}
\title{Calculating the correlation between the distance matrixes of parasites and their associated hosts}
\usage{
get.PHdistCorrelation(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite tree}
}
\description{
Calculating the correlation between the distance matrixes of parasites and their associated hosts
}
\examples{
get.PHdistCorrelation()
}
\keyword{correlation}
\keyword{distance,}
\keyword{genetic}

