% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_GDist}
\alias{get_GDist}
\title{Calculating the distance matrix between host branches alive at a particular timepoint}
\usage{
get_GDist(branches, t = NA)
}
\arguments{
\item{branches}{raw branches matrix of a host tree}

\item{t}{timepoint in simulation at which you want distance information}
}
\description{
The following function returns a matrix of genetic distances between all species present at a given time. If time t is not specified, the end point of the tree is used.
}
\keyword{internal}
