% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_hostShifts}
\alias{get_hostShifts}
\title{Counting parasite host-jumps.}
\usage{
get_hostShifts(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one host and one parasite tree.}
}
\description{
The following function counts the host-jumps of a host-parasite phylogeny
}
\keyword{cophylogeny,}
\keyword{host-jumps}
\keyword{internal}
