% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_infectionFrequenciesSubtrees}
\alias{get_infectionFrequenciesSubtrees}
\title{The following function returns the fraction of infected host species within a
subclade of a host tree that is specified by tips, a vector of tip labels.}
\usage{
get_infectionFrequenciesSubtrees(cophy, tips)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one host
and one parasite tree.}

\item{tips}{a vector of tip labels}
}
\description{
The following function returns the fraction of infected host species within a
subclade of a host tree that is specified by tips, a vector of tip labels.
}
\examples{
get_infectionFrequenciesSubtrees
}
\keyword{frequency}
\keyword{infection}
\keyword{subtree,}
