% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_infectionStatistics}
\alias{get_infectionStatistics}
\title{Basic statistics describing a cophylogeny.}
\usage{
get_infectionStatistics(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one host
and one parasite tree.}
}
\description{
The following function calculates the basic statistics concerning a
particular simulation including: number of extant host species, number of
extant parasite species, percentage of host species that are infected by at
least one parasite, mean number of parasites per host
}
\examples{
cop<-rcophylo(tmax=5)
get_infectionStatistics(cophy=cop)
}
\keyword{cophylogeny,}
\keyword{statistics}
