% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nodeTime}
\alias{get_nodeTime}
\title{Function to obtain the time of a node relative to the root}
\usage{
get_nodeTime(phy, node)
}
\arguments{
\item{phy}{tree in phylo format}

\item{node}{a particular node in the tree}
}
\description{
Function to obtain the time of a node relative to the root
}
\keyword{node,}
\keyword{root}
