% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_PextinctionTime}
\alias{get_PextinctionTime}
\title{Parasite lineage extinction time}
\usage{
get_PextinctionTime(x)
}
\arguments{
\item{x}{either a cophylogeny (object of class "cophylogeny") containing one host
and one parasite tree, or just a parasite tree (of class "phylo")}
}
\description{
Function to calculate the time of the last surviving parasite species
}
\examples{
coph <- rcophylo(tmax=5)
get_PextinctionTime(coph)
}
