% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_PHDistSubtreeCorrelation}
\alias{get_PHDistSubtreeCorrelation}
\title{Calculating the correlation between the distance matrixes of parasites and
their associated hosts within subtrees specified by particular height.
Requires at least three living parasites.}
\usage{
get_PHDistSubtreeCorrelation(cophy, h = NULL, k = NULL)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one
host and one parasite tree.}

\item{h}{numeric scalar or vector with heights where the tree should be cut.}

\item{k}{an integer scalar or vector with the desired number of groups}
}
\description{
Calculating the correlation between the distance matrixes of parasites and
their associated hosts within subtrees specified by particular height.
Requires at least three living parasites.
}
\examples{
HPtree<-rcophylo_HP(tmax=10, K=10)
get_PHDistSubtreeCorrelation(cophy=HPtree, h=2)
}
\keyword{correlation}
\keyword{distance,}
\keyword{genetic}
