% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tBirth}
\alias{get_tBirth}
\title{A (recursive) function to obtain the time of birth of a branch n, given a tree
in phylo format and a list of ancestor branches for that tree}
\usage{
get_tBirth(n, root.edge, edge.length, ancBranches)
}
\arguments{
\item{n}{some branch in a tree}

\item{root.edge}{object sourced from tree in phylo format}

\item{edge.length}{object sourced from tree in phylo format}

\item{ancBranches}{the ancestral branches for the tree}
}
\description{
A (recursive) function to obtain the time of birth of a branch n, given a tree
in phylo format and a list of ancestor branches for that tree
}
