% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{nodeTime}
\alias{nodeTime}
\title{Function to obtain the time of a node relative to the root}
\usage{
nodeTime(phy, node)
}
\arguments{
\item{phy:}{tree in phylo format}

\item{node:}{a particular node in the tree}
}
\description{
Function to obtain the time of a node relative to the root
}
\examples{
nodeTime()
}
\keyword{node,}
\keyword{root}

