% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{parsimulate.2PonH.singleparam}
\alias{parsimulate.2PonH.singleparam}
\title{A function to simulate many random coevolving (dual)parasite phylogenies on pre-built host-trees and calculate statistics in parallel.}
\usage{
parsimulate.2PonH.singleparam(Htrees, fromHtree = NA, toHtree = NA, tmax,
  P.startT, beta, gamma.P, gamma.Q, sigma.self, sigma.cross, mu.P, mu.Q,
  timestep, reps1, reps2, filename = NA, ncores)
}
\arguments{
\item{Htrees:}{pre-built host trees on which to simulate parasite trees, fromHtree: starting host-tree, toHtree: finishing host-tree, tmax: maximum time for which to simulate, P.startT: the timepoint at which a parasite invades the host-tree, beta: parasite host jump rate, gamma.P: dependency on genetic distance for host jumps, gamma.Q: dependency on genetic distance for host jumps, sigma.self: probability of successful co-infection with related parasite following host jump, sigma.cross: probability of successful co-infection with unrelated parasite following host jump, mu.P: parasite extinction rate, timestep: timestep for simulations, mu.Q: parasite extinction rate, timestep: timestep for simulations, reps1: the number of starting points for the parasite trees, reps2: the number of replicates per starting point, filename: name underwhich set of simulations and statistics will be saved, ncores: the number of cores that will be used to run simulations in parallel.}
}
\description{
The following function simulates two competing parasite phylogenetic trees on pre-built host trees using parallel computing.
}
\examples{
parsimulate.2PonH.singleparam()
}
\keyword{Host-Parasite}
\keyword{multiple}
\keyword{parallel}
\keyword{phylogeny,}
\keyword{statistics,}

