% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{parsimulate.PonH.singleparam}
\alias{parsimulate.PonH.singleparam}
\title{A function to simulate many random coevolving parasite phylogenies on pre-built host-trees and calculate statistics in parallel.}
\usage{
parsimulate.PonH.singleparam(Htrees, fromHtree = NA, toHtree = NA, tmax,
  P.startT, beta, gamma, sigma, muP, timestep, reps1, reps2, filename = NA,
  ncores)
}
\arguments{
\item{Htrees:}{pre-built host trees on which to simulate parasite trees}

\item{fromHtree:}{starting host-tree}

\item{toHtree:}{finishing host-tree}

\item{tmax:}{maximum time for which to simulate}

\item{P.startT:}{the timepoint at which a parasite invades the host-tree}

\item{beta:}{parasite host jump rate}

\item{gamma:}{dependency on genetic distance for host jumps}

\item{sigma:}{probability of successful co-infection following host jump}

\item{muP:}{parasite extinction rate}

\item{timestep:}{timestep for simulations}

\item{reps1:}{the number of starting points for the parasite trees}

\item{reps2:}{the number of replicates per starting point}

\item{filename:}{name underwhich set of simulations and statistics will be saved}

\item{ncores:}{the number of cores that will be used to run simulations in parallel}
}
\description{
The following function simulates parasite phylogenetic trees on pre-built host trees using parallel computing.
}
\examples{
parsimulate.PonH.singleparam()
}
\keyword{Host-Parasite}
\keyword{multiple}
\keyword{parallel}
\keyword{phylogeny,}
\keyword{statistics,}

