% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{plot.2Pcophy}
\alias{plot.2Pcophy}
\title{The following function plots a host and 2 parasite phylogenies}
\usage{
\method{plot}{2Pcophy}(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and two parasite trees}
}
\description{
The following function plots a host-(dual)parasite phylogeny
}
\examples{
plot.2Pcophy()
}
\keyword{cophylogeny,}
\keyword{plot}

