% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{plot.cophy}
\alias{plot.cophy}
\title{The following function plots both a host and parasite phylogeny}
\usage{
\method{plot}{cophy}(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (in phylo format) containing one host and one parasite tree}
}
\description{
The following function plots a host-parasite phylogeny
}
\examples{
plot.cophy()
}
\keyword{cophylogeny,}
\keyword{plot}
