% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_resistance}
\alias{plot_resistance}
\title{The following function plots a host phylogeny including the evolution of the resistance trait}
\usage{
plot_resistance(Hphy, TraitTracking)
}
\arguments{
\item{Hphy:}{a phylogeny (in phylo format) containing one host tree}
}
\description{
The following function plots a host-parasite phylogeny
}
\examples{
plot.resistance()
}
\keyword{phylogeny,}
\keyword{plot,}
\keyword{resistance}
