% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prune_hostTree}
\alias{prune_hostTree}
\title{A function to prune a host tree according to some bias}
\usage{
prune_hostTree(Htree, propSampled = 1, bias = 0)
}
\arguments{
\item{Htree}{the host tree to be pruned}
}
\description{
A function to prune a host tree according to some bias
}
\examples{
Htree <- rphylo_H(tmax=5)
prune_hostTree(Htree)
}
