% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prune_parasiteTree}
\alias{prune_parasiteTree}
\title{A function to prune a host tree according to some bias}
\usage{
prune_parasiteTree(Htree, Ptree)
}
\arguments{
\item{Htree}{a host tree}

\item{Ptree}{a parasite tree to be pruned}
}
\description{
A function to prune a host tree according to some bias
}
\examples{
HPtree <- rcophylo_HP(5)
prune_parasiteTree(Htree = HPtree[[1]], Ptree = HPtree[[2]])
}
