% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{randomcophy.HP}
\alias{randomcophy.HP}
\title{A random cophylogeny building function}
\usage{
randomcophy.HP(tmax, nHmax = Inf, lambda = 1, muH = 0.5, K = Inf,
  beta = 0.1, gamma = 0.2, sigma = 0, muP = 0.5,
  prune.extinct = FALSE, export.format = "Phylo", timestep = 0.001)
}
\arguments{
\item{tmax:}{maximum time for which to simulate, nHmax: maximum host species number until which to simulate, lambda: host speciation rate, K: carrying capacity for host species, muH: host extinction rate, beta: parasite host jump rate, gamma: dependency on genetic distance for host jumps, sigma: probability of successful co-infection following host jump, muP: parasite extinction rate, prune.extinct: whether to remove all extinct branches defaulting to FALSE, export.format: either "Phylo" (exported in Ape Phylo format, the default setting)) or "Raw" (just a list of branches as used within the function itself), timestep: timestep for simulations}
}
\description{
The following function simulates a host phylogenetic tree on which a parasite clade co-diversifies.
}
\examples{
randomcophy.HP()
}
\keyword{Host-Parasite}
\keyword{phylogeny}

