% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{randomcophy.PonH}
\alias{randomcophy.PonH}
\title{A random parasite tree building function}
\usage{
randomcophy.PonH(tmax, H.tree, beta = 0.1, gamma = 0.2, sigma = 0,
  muP = 0.5, prune.extinct = FALSE, export.format = "Phylo",
  P.startT = 0, ini.Hbranch = NA, Gdist = NA, timestep = 0.001)
}
\arguments{
\item{tmax:}{maximum time for which to simulate}

\item{H.tree:}{a pre-built host phylogenetic tree}

\item{beta:}{parasite host jump rate}

\item{gamma:}{dependency on genetic distance for host jumps}

\item{sigma:}{probability of successful co-infection following host jump}

\item{muP:}{parasite extinction rate}

\item{prune.extinct:}{whether to remove all extinct branches defaulting to FALSE}

\item{export.format:}{either "Phylo" (exported in Ape Phylo format, the default setting)) or "Raw" (just a list of branches as used within the function itself)}

\item{P.startT:}{the timepoint at which a parasite invades the host-tree}

\item{ini.Hbranch:}{the host branch from which the parasite invasion is initiated (defaults to NA)}

\item{Gdist:}{can input a pre-calculated distance matrix of the living host branches at time of infection (defaults to NA)}

\item{timestep:}{timestep for simulations}
}
\description{
The following function simulates a parasite phylogenetic tree on a pre-built host phylogeny.
}
\examples{
randomcophy.PonH()
}
\keyword{Host-Parasite}
\keyword{phylogeny}

