% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{randomphy.H}
\alias{randomphy.H}
\title{A random host tree building function}
\usage{
randomphy.H(tmax, nHmax = Inf, lambda = 1, muH = 0.5, K = Inf,
  prune.extinct = FALSE, export.format = "Phylo", timestep = 0.001)
}
\arguments{
\item{tmax:}{maximum time for which to simulate, nHmax: maximum host species number until which to simulate, lambda: host speciation rate, K: carrying capacity for host species, muH: host extinction rate, prune.extinct: whether to remove all extinct branches defaulting to FALSE, export.format: either "Phylo" (exported in Ape Phylo format, the default setting) or "Raw" (just a list of branches as used within the function itself), timestep: timestep for simulations}
}
\description{
The following function simulates a host phylogenetic tree.
}
\examples{
randomphy.H()
}
\keyword{Host}
\keyword{phylogeny}

