% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcophylo.R
\name{rcophylo_PonH}
\alias{rcophylo_PonH}
\title{Cophylogeny simulation on an existing host tree.}
\usage{
rcophylo_PonH(H.tree, beta = 0.1, gamma = 0.2, sigma = 0, nu = 0.5,
  prune.extinct = FALSE, export.format = "Phylo", P.startT = 0,
  ini.Hbranch = NA, Gdist = NA, timestep = 0.001)
}
\arguments{
\item{H.tree}{a pre-built host phylogenetic tree.}

\item{beta}{parasite host jump rate.}

\item{gamma}{a numeric value giving the dependency of host shift success of a parasite on phylogenetic distance between the old and the new host.}

\item{sigma}{a numeric value determining how successful host shift are when the new host is already infected by a parasite. 
Specifically, the probability of host shift success \eqn{(1-\sigma)^n}, 
where \eqn{n} is the number of pre-existing parasites on the new host branch.}

\item{nu}{parasite extinction rate.}

\item{prune.extinct}{logical. Determines whether or not to remove all extinct branches.}

\item{export.format}{either "Phylo" (exported in Ape Phylo format, the default setting)) or "Raw" (a matrix where rows are all the branches, this is the used format internally).}

\item{P.startT}{the timepoint at which a parasite invades the host tree.}

\item{ini.Hbranch}{numerical. The host branch number from which the parasite invasion is initiated. If left at the default value NA, a randomly chosen branch will be selected.}

\item{Gdist}{optional: a pre-calculated distance matrix of the living host branches at time of infection. 
Providing this matrix will speed up the calculation which may be useful when running several simulations on the same host tree.}

\item{timestep}{a numeric value giving the time step by which the simulation proceeds. 
Increase to make the simulation faster or decrease to make it more precise.}
}
\value{
By default, an object of class "cophylo" is returned that is a list of phylo objects (one for the host and one for the parasite), as specified in the R-package "ape". 
  If the argument \code{export.format} is set to "Raw" the function returns a list of dataframes containing information on all the branches in the trees. 
  (These dataframes are what the function uses internally.)
}
\description{
This function simulates the codiversification of a clade of parasites on a given host phylogeny (simulated or estimated) that is provided.
}
\examples{
rcophylo_PonH()
}
\keyword{Host-Parasite}
\keyword{phylogeny}
