% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{simulate.PonH.singleparam}
\alias{simulate.PonH.singleparam}
\title{A function to simulate many random parasite phylogenies on pre-built host-trees and calculate statistics}
\usage{
\method{simulate}{PonH.singleparam}(Htrees, fromHtree = NA, toHtree = NA,
  tmax, P.startT, beta, gamma, sigma, muP, timestep, reps1 = 1, reps2,
  filename = NA)
}
\arguments{
\item{Htrees:}{pre-built host trees on which to simulate parasite trees}

\item{fromHtree:}{starting host-tree}

\item{toHtree:}{finishing host-tree}

\item{tmax:}{maximum time for which to simulate}

\item{P.startT:}{the timepoint at which a parasite invades the host-tree}

\item{beta:}{parasite host jump rate}

\item{gamma:}{dependency on genetic distance for host jumps}

\item{sigma:}{probability of successful co-infection following host jump}

\item{muP:}{parasite extinction rate}

\item{timestep:}{timestep for simulations}

\item{reps1:}{the number of starting points for the parasite trees}

\item{reps2:}{the number of replicates per starting point}

\item{filename:}{name underwhich set of simulations and statistics will be saved}
}
\description{
A function to run a certain number of replicate simulations, save all the trees and output all stats.
}
\examples{
simulate.PonH.singleparam()
}
\keyword{Host-Parasite}
\keyword{multiple}
\keyword{phylogeny,}
\keyword{statistics}

