% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions v0.928 17-01-29.R
\name{simulate.singleparam}
\alias{simulate.singleparam}
\title{A function to simulate many random cophylogenies and calculate statistics}
\usage{
\method{simulate}{singleparam}(tmax = 0, lambda, muH, beta, gamma, sigma, muP,
  K, timestep, reps, filename = NA)
}
\arguments{
\item{tmax:}{maximum time for which to simulate, lambda: host speciation rate, K: carrying capacity for host species, muH: host extinction rate, beta: parasite host jump rate, gamma: dependency on genetic distance for host jumps, sigma: probability of successful co-infection following host jump, muP: parasite extinction rate, timestep: timestep for simulations, reps: number of times to simulate this set of parameters, filename: name underwhich set of simulations and statistics will be saved}
}
\description{
A function to run a certain number of replicate simulations, save all the trees and output all stats.
}
\examples{
simulate.singleparam()
}
\keyword{Host-Parasite}
\keyword{phylogeny,}
\keyword{statistics}

