% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cophys_H}
\alias{simulate_cophys_H}
\title{A function to simulate many random host phylogenies}
\usage{
simulate_cophys_H(tmax = 0, lambda, mu, K, timestep, reps, filename = NA)
}
\arguments{
\item{tmax:}{maximum time for which to simulate}

\item{lambda:}{host speciation rate}

\item{K:}{carrying capacity for host species}

\item{mu:}{host extinction rate}

\item{timestep:}{timestep for simulations}

\item{reps:}{number of times to simulate this set of parameters}

\item{filename:}{name underwhich set of simulations and statistics will be saved}
}
\description{
A function to run a certain number of replicate simulations, save all the trees and output all stats.
}
\examples{
simulate_cophys_H()
}
\keyword{Host}
\keyword{multiple}
\keyword{phylogenies}
