% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cophys_PonH_Htrait}
\alias{simulate_cophys_PonH_Htrait}
\title{A parallel parasite tree building function with host response to infection}
\usage{
simulate_cophys_PonH_Htrait(Htrees, HtreesPhylo = NA, fromHtree = NA,
  toHtree = NA, beta = 0.1, gamma = 0.2, sigma = 0, nu = 0.5, kappa,
  delta, epsilon.1to0, epsilon.0to1, startTrait, omega, rho, psi,
  TraitTracking = NA, prune.extinct = FALSE, export.format = "Phylo",
  P.startT = 0, reps1 = 1, reps2 = 1, ini.Hbranch = NA, Gdist = NA,
  timestep = 0.001, filename = NA, ncores = 1)
}
\arguments{
\item{Htrees}{a pre-built host phylogenetic tree}

\item{HtreesPhylo}{list of host trees converted to phylo format}

\item{fromHtree}{starting host-tree}

\item{toHtree}{finishing host-tree}

\item{beta}{parasite host jump rate}

\item{gamma}{dependency on genetic distance for host jumps}

\item{sigma}{probability of successful co-infection following host jump}

\item{nu}{parasite extinction rate}

\item{kappa}{the probability that a parasite duplicates within a host species}

\item{delta}{the probability that a parasite is lost randomly in one of the daughter host species during a cospeciation event}

\item{epsilon.1to0}{the baseline rate that a host with trait value 1 will mutate to a host with trait value 0}

\item{epsilon.0to1}{the baseline rate that a host with trait value 0 will mutate to a host with trait value 1}

\item{startTrait}{specifies the initial resistance trait of the first host species (0 or 1). Defaults to NA (random)}

\item{omega}{factor by which switching between trait values is altered depending on the trait value of the host and presence of parasites}

\item{rho}{factor by which parasite extinction rate increases in response to host resistance}

\item{psi}{factor by which parasite host-jump success decreases due to resistance of the new host}

\item{TraitTracking}{an object which provides the evolutionary history of the parasite interaction trait. Needs to match host trees provided.}

\item{prune.extinct}{whether to remove all extinct branches defaulting to FALSE}

\item{export.format}{either "Phylo" (exported in Ape Phylo format, the default setting)) or "Raw" (just a list of branches as used within the function itself)}

\item{P.startT}{the timepoint at which a parasite invades the host-tree}

\item{reps1}{the number of starting points for the parasite trees}

\item{reps2}{the number of replicates per starting point}

\item{ini.Hbranch}{the host branch from which the parasite invasion is initiated (defaults to NA)}

\item{Gdist}{can input a pre-calculated distance matrix of the living host branches at time of infection (defaults to NA)}

\item{timestep}{timestep for simulations}

\item{filename}{name underwhich set of simulations and statistics will be saved}

\item{ncores}{the number of cores that will be used to run simulations in parallel}
}
\description{
The following function simulates a parasite phylogenetic tree on a pre-built host phylogeny.
}
\examples{
Htree<-list(rphylo_H(tmax=5, export.format="Raw"), rphylo_H(tmax=5, export.format="Raw"))
simulate_cophys_PonH_Htrait(Htrees=Htree, P.startT=2.5, beta=0.1, gamma=0.02, sigma=0, nu=0.5,
kappa=0, delta=0, epsilon.1to0=0.001, epsilon.0to1=0.001, startTrait=0, omega=10, rho=0.75,
psi=0.75, timestep=0.001, filename="file")
}
\keyword{Host-Parasite}
\keyword{phylogeny}
