% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cophys_PQonH}
\alias{simulate_cophys_PQonH}
\title{A function to simulate many random coevolving (dual)parasite phylogenies on pre-built host-trees and calculate statistics in parallel.}
\usage{
simulate_cophys_PQonH(Htrees, fromHtree = NA, toHtree = NA, P.startT, beta,
  gamma.P, gamma.Q, sigma.self, sigma.cross, nu.P, nu.Q, kappa.P, kappa.Q,
  delta.P, delta.Q, timestep, reps1 = 1, reps2 = 1, filename = NA,
  ncores = 1, DBINC = 100)
}
\arguments{
\item{Htrees:}{pre-built host trees on which to simulate parasite trees}

\item{fromHtree:}{starting host-tree}

\item{toHtree:}{finishing host-tree}

\item{P.startT:}{the timepoint at which a parasite invades the host-tree}

\item{beta:}{parasite host jump rate}

\item{gamma.P:}{dependency on genetic distance for host jumps}

\item{gamma.Q:}{dependency on genetic distance for host jumps}

\item{sigma.self:}{probability of successful co-infection with related parasite following host jump}

\item{sigma.cross:}{probability of successful co-infection with unrelated parasite following host jump}

\item{nu.P:}{parasite extinction rate}

\item{timestep:}{timestep for simulations}

\item{nu.Q:}{parasite extinction rate}

\item{timestep:}{timestep for simulations}

\item{reps1:}{the number of starting points for the parasite trees}

\item{reps2:}{the number of replicates per starting point}

\item{filename:}{name underwhich set of simulations and statistics will be saved}

\item{ncores:}{the number of cores that will be used to run simulations in parallel}
}
\description{
The following function simulates two competing parasite phylogenetic trees on pre-built host trees using parallel computing.
}
\examples{
simulate_cophys_PQonH()
}
\keyword{Host-Parasite}
\keyword{multiple}
\keyword{parallel}
\keyword{phylogeny,}
\keyword{statistics,}
