% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_HostTrees}
\alias{simulate_HostTrees}
\title{A function to simulate many random host phylogenies}
\usage{
simulate_HostTrees(tmax, lambda, mu, K, timestep, reps, filename = NA)
}
\arguments{
\item{tmax}{maximum time for which to simulate}

\item{lambda}{host speciation rate}

\item{mu}{host extinction rate}

\item{K}{carrying capacity for host species}

\item{timestep}{timestep for simulations}

\item{reps}{number of times to simulate this set of parameters}

\item{filename}{name underwhich set of simulations and statistics will be saved}
}
\description{
A function to run a certain number of replicate simulations, save all the trees and output all stats.
}
\examples{
simulate_HostTrees(tmax=5,lambda=1,mu=0.5,K=5,timestep=0.001,reps=10,filename="file")
}
\keyword{Host}
\keyword{multiple}
\keyword{phylogenies}
