% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{subtree.freqInfected}
\alias{subtree.freqInfected}
\title{The following function returns the fraction of infected host species within a subclade of a host tree that is specified by tips, a vector of tip labels.}
\usage{
subtree.freqInfected(cophy, tips)
}
\arguments{
\item{cophy:}{cophylogeny in phylo format}

\item{tips:}{a vector of tip labels}
}
\description{
The following function returns the fraction of infected host species within a subclade of a host tree that is specified by tips, a vector of tip labels.
}
\examples{
subtree.freqInfected()
}
\keyword{frequency}
\keyword{infection}
\keyword{subtree,}
