% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cophylogeny}
\alias{cophylogeny}
\title{Creates a cophy object}
\usage{
cophylogeny(H.tree, P.tree, Q.tree = NA, TraitTracking = NA)
}
\arguments{
\item{H.tree}{a pre-built host phylogenetic tree of class 'phylo' (required)}

\item{P.tree}{a pre-built parasite phylogenetic tree of class 'phylo'
(required)}

\item{Q.tree}{a pre-built parasite phylogenetic tree of class 'phylo'
(optional)}

\item{TraitTracking}{an object that tracks the evolution of a parasite
response trait on the host tree (optional)}
}
\description{
This function creates an object of class 'cophy', which can be passed to the
plot.cophy() function. This object must contain at least one host and one
parasite tree, but can also contain a second parasite tree and a
TraitTracking object.
}
\examples{
Htree<-rphylo_H(tmax=5, export.format="Raw")
HPtree<-rcophylo_PonH(H.tree=Htree, tmax=5)
cophylogeny(HPtree[[1]], HPtree[[2]])
}
\keyword{cophy,}
\keyword{object}
