% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_HostShifts}
\alias{get_HostShifts}
\title{Counting parasite host-jumps.}
\usage{
get_HostShifts(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one host and one parasite tree.}
}
\description{
The following function counts the host-jumps of a host-parasite phylogeny
}
\examples{
HPtree<-rcophylo_HP(tmax=5)
get_HostShifts(cophy=HPtree)
}
\keyword{cophylogeny,}
\keyword{host-jumps}
