% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_infectionFrequencies}
\alias{get_infectionFrequencies}
\title{The following function counts infection levels}
\usage{
get_infectionFrequencies(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylo") containing one host and one parasite tree.}
}
\description{
The following function counts the number of hosts infected with a particular number of parasites from the same parasite phylogeny
}
\examples{
get_infectionFrequencies()
}
\keyword{cophylogeny,}
\keyword{infection-counts}
