% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_infectionFrequencies}
\alias{get_infectionFrequencies}
\title{Distribution of parasite numbers on the tips of a cophylogeny.}
\usage{
get_infectionFrequencies(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one host
and one parasite tree.}
}
\description{
This function returns the numbers of extant hosts infected with a particular
number of parasites (0, 1, 2, ...).
}
