% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_PextinctionTime}
\alias{get_PextinctionTime}
\title{Function to calculate the time of the last surviving parasite species}
\usage{
get_PextinctionTime(phy)
}
\arguments{
\item{phy}{parasite tree in phylo format}
}
\description{
Function to calculate the time of the last surviving parasite species
}
\examples{
HPtree<-rcophylo_HP(tmax=5)
get_PextinctionTime(phy=HPtree[[2]])
}
\keyword{last}
\keyword{parasite}
