% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_PHDistCorrelation}
\alias{get_PHDistCorrelation}
\title{Correlation between pairwise distances of associated hosts and parasites.}
\usage{
get_PHDistCorrelation(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylogeny") containing one
host and one parasite tree.}
}
\description{
Calculating the correlation between the distance matrixes of parasites and
their associated hosts
}
\examples{
HPtree<-rcophylo_HP(tmax=5)
get_PHDistCorrelation(cophy=HPtree)
}
\keyword{correlation}
\keyword{distance,}
\keyword{genetic}
