% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_PHDistCorrelation}
\alias{get_PHDistCorrelation}
\title{Calculating the correlation between the distance matrixes of parasites and their associated hosts}
\usage{
get_PHDistCorrelation(cophy)
}
\arguments{
\item{cophy}{a cophylogeny (object of class "cophylo") containing one host and one parasite tree.}
}
\description{
Calculating the correlation between the distance matrixes of parasites and their associated hosts
}
\examples{
get_PHDistCorrelation()
}
\keyword{correlation}
\keyword{distance,}
\keyword{genetic}
