% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tBirth}
\alias{get_tBirth}
\title{(recursive) function to obtain the time of birth of a branch n, given a tree in phy in phylo format and a list of ancestor branches for that tree}
\usage{
get_tBirth(n, root.edge, edge.length, ancBranches)
}
\arguments{
\item{n:}{some branch in a tree}

\item{phy:}{the tree in phylo format}

\item{ancBranches:}{the ancestral branches for the tree}
}
\description{
(recursive) function to obtain the time of birth of a branch n, given a tree in phy in phylo format and a list of ancestor branches for that tree
}
