% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cophylo}
\alias{plot.cophylo}
\title{Cophylogeny plot}
\usage{
\method{plot}{cophylo}(cophy, ParasiteCol = c("Red", "Blue"),
  TraitTracking = NA, ResistanceCol = "lawn green")
}
\arguments{
\item{cophy:}{a cophylogeny (object of class "cophylo") containing a host tree and a parasite tree. This may also contain either a second parasite tree or host and parasite trait values.}

\item{ParasiteCol:}{a list of length 2 that specifies the colours to use when ploting parasite lineages. The first position indicates the colour of the first parasite lineage. If there is a second parasite lineage, its colour is specified by the second position. Defaults to red and blue.}

\item{TraitTracking:}{an object that tracks the evolution of host resistance traits if available.}

\item{ResistanceCol:}{in the case that a TraitTracking object is available, gives the option to choose the plotting colour.}
}
\description{
This function plots a host-parasite cophylogenetic tree,
}
\examples{
plot.cophylo()
}
\keyword{cophylogeny,}
\keyword{plot}
