% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cophylo}
\alias{plot.cophylo}
\title{Cophylogeny plot}
\usage{
\method{plot}{cophylo}(cophy)
}
\arguments{
\item{cophy:}{a cophylogeny (object of class "cophylo") containing a host tree and a parasite tree. This may also contain either a second parasite tree or host and parasite trait values.}
}
\description{
This function plots a host-parasite cophylogenetic tree,
}
\examples{
plot.cophylo()
}
\keyword{cophylogeny,}
\keyword{plot}
