% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cophylogeny}
\alias{plot.cophylogeny}
\title{Cophylogeny plot}
\usage{
\method{plot}{cophylogeny}(x, ParasiteCol = "Red", ...)
}
\arguments{
\item{x}{a cophylogeny of class 'cophylogeny', containing a host tree and a
parasite tree.}

\item{ParasiteCol}{specifies the colour to use when ploting parasite lineages.
Defaults to "Red".}

\item{...}{other parameters to be passed to plotting functions.}
}
\description{
This function plots a host-parasite cophylogenetic tree.
}
\examples{
Htree<-rphylo_H(tmax=5, export.format='raw')
HPtree<-rcophylo_PonH(H.tree=Htree)
plot(cophylogeny(HPtree))
}
\keyword{cophylogeny,}
\keyword{plot}
