% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cophylogeny}
\alias{plot.cophylogeny}
\title{Cophylogeny plot}
\usage{
\method{plot}{cophylogeny}(x, hostCol = "Black", parasiteCol = "Red",
  ...)
}
\arguments{
\item{x}{a cophylogeny of class 'cophylogeny', containing a host tree and a
parasite tree.}

\item{hostCol}{specifies the colour to use when ploting host lineages.
Defaults to "Black".}

\item{parasiteCol}{specifies the colour to use when ploting parasite lineages.
Defaults to "Red".}

\item{...}{other parameters to be passed to plotting functions.}
}
\description{
This function plots a host-parasite cophylogenetic tree.
}
\examples{
Htree<-rphylo_H(tmax=5, exportFormat='raw')
HPtree<-rcophylo(HTree=Htree)
plot(cophylogeny(HPtree))
}
\keyword{cophylogeny,}
\keyword{plot}
