% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.cophylogeny}
\alias{print.cophylogeny}
\title{Printing a cophylogeny}
\usage{
\method{print}{cophylogeny}(x, ...)
}
\arguments{
\item{x}{object of class "cophylogeny"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This functions prints information about a cophylogeny,
listing the general structure of the cophylogeny
along with properties of the host and parasite trees it consists of.
}
\examples{
cop<-rcophylo_HP(tmax=5, K=5)
print(cop)
}
