% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prune_cophylo}
\alias{prune_cophylo}
\title{A function that prunes a cophylogeny and plots the resulting co-cladogram}
\usage{
prune_cophylo(cophy)
}
\arguments{
\item{cophy}{an object of class cophylogeny that contains a host tree with one
associated parasite tree.}
}
\description{
A function that prunes a cophylogeny and plots the resulting co-cladogram
}
\examples{
coph <- rcophylo(tmax=5)
prune_cophylo(coph)
}
