% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prune_Cophylo}
\alias{prune_Cophylo}
\title{A function that prunes a cophylogeny and plots the resulting co-cladogram}
\usage{
prune_Cophylo(cophy, plot = TRUE)
}
\arguments{
\item{cophy}{an object of class cophylogeny that contains a host tree with one
associated parasite tree.}

\item{plot}{logical. Whether or not the output should automatically be plotted.}
}
\description{
A function that prunes a cophylogeny and plots the resulting co-cladogram
}
\examples{
cophy <- rcophylo_HP(5)
prune_Cophylo(cophy)
}
