% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcophylo.R
\name{rcophylo_HP}
\alias{rcophylo_HP}
\title{Cophylogeny simulation}
\usage{
rcophylo_HP(
  tmax,
  nHmax = Inf,
  lambda = 1,
  mu = 0.5,
  K = Inf,
  beta = 0.1,
  gamma = 0.02,
  sigma = 0,
  nu = 0.5,
  kappa = 0,
  delta = 0,
  exportFormat = "cophylogeny",
  timestep = 0.001
)
}
\arguments{
\item{tmax}{a numeric value giving the length of the simulation.}

\item{nHmax}{a numeric value giving the number of host species at which the
simulation is stopped.}

\item{lambda}{a numeric value giving the host speciation rate.}

\item{mu}{a numeric value giving the host extinction rate.}

\item{K}{a numeric value giving the carrying capacity for the host species.}

\item{beta}{a numeric value giving the baseline parasite host shift rate.}

\item{gamma}{a numeric value giving the dependency of host shift success on
phylogenetic distance between the old and the new host.}

\item{sigma}{a numeric value determining how successful host shift are when
the new host is already infected by a parasite. Specifically, the
probability of host shift success \eqn{(1-\sigma)^n}, where \eqn{n} is the
number of pre-existing parasites on the new host branch.}

\item{nu}{a numeric value giving the parasite extinction rate.}

\item{kappa}{a numeric value giving the parasite speciation rate within
hosts.}

\item{delta}{a numeric value giving the probability of lineage loss during
co-speciation. delta=0 specifies faithful transmission of the parasites to
both new host species, whereas delta=1 specifies that parasites will only
be inherited by one daughter host species.}

\item{exportFormat}{a string specifying either "cophylogeny" or "raw". Where
"cophylogeny" specifies that the output will be exported as class
"cophylogeny" (the default). "raw" exports the output as two objects of
class "data.frame".}

\item{timestep}{a numeric value giving the time step by which the simulation
proceeds. Increase to make the simulation faster or decrease to make it
more precise.}
}
\value{
By default, an object of class "cophylogeny" is returned that is a
  list of phylo objects (one for the host and one for the parasite), as
  specified in the R-package "ape". If the argument \code{exportFormat} is
  set to "raw" the function returns a list of dataframes containing
  information on all the branches in the trees. (These dataframes are what
  the function uses internally.)
}
\description{
This function simulates the evolution of a random host tree with a coevolving
random parasite tree.
}
\keyword{internal}
