% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcophylo.R
\name{rcophylo_PonH}
\alias{rcophylo_PonH}
\title{Cophylogeny simulation on an existing host tree.}
\usage{
rcophylo_PonH(HTree, beta = 0.1, gamma = 0, sigma = 0, nu = 0.5,
  kappa = 0, delta = 0, exportFormat = "cophylogeny", PStartT = 0,
  iniHBranch = NULL, Gdist = NULL, timestep = 0.001)
}
\arguments{
\item{HTree}{a pre-built host phylogenetic tree.}

\item{beta}{a numeric value giving the baseline parasite host shift rate.}

\item{gamma}{a numeric value giving the dependency of host shift success of a
parasite on phylogenetic distance between the old and the new host.}

\item{sigma}{a numeric value determining how successful host shift are when
the new host is already infected by a parasite. Specifically, the
probability of host shift success \eqn{(1-\sigma)^n}, where \eqn{n} is the
number of pre-existing parasites on the new host branch.}

\item{nu}{a numeric value giving the parasite extinction rate.}

\item{kappa}{a numeric value giving the parasite speciation rate within
hosts.}

\item{delta}{a numeric value giving the probability of lineage loss during
co-speciation. delta=0 specifies faithful transmission of the parasites to
both new host species, whereas delta=1 specifies that parasites will only
be inherited by one daughter host species.}

\item{exportFormat}{a string specifying either "cophylogeny" or "raw". Where
"cophylogeny" specifies that the output will be exported as class
"cophylogeny" (the default). "raw" exports the output as two objects of
class "data.frame".}

\item{PStartT}{the timepoint at which a parasite invades the host tree
(default set to 0).}

\item{iniHBranch}{numerical. The host branch number from which the parasite
invasion is initiated. If left at the default value NA, a randomly chosen
host branch alive at PStartT (time of infection) will be selected.}

\item{Gdist}{optional: a pre-calculated distance matrix of the living host
branches at PStartT (time of infection). Providing this matrix will speed
up the calculation which may be useful when running several simulations on
the same host tree.}

\item{timestep}{a numeric value giving the time step by which the simulation
proceeds. Increase to make the simulation faster or decrease to make it
more precise. The same timestep that was used to build the host tree,
should be used to build the parasite tree to avoid potential errors.}
}
\value{
By default, an object of class "cophylogeny" is returned that is a
  list of phylo objects (one for the host and one for the parasite), as
  specified in the R-package "ape". If the argument \code{exportFormat} is
  set to "raw" the function returns a list of dataframes containing
  information on all the branches in the trees. (These dataframes are what
  the function uses internally.)
}
\description{
This function simulates the codiversification of a clade of parasites on a
given host phylogeny (simulated or estimated) that is provided.
}
\keyword{internal}
