% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophy functions.R
\name{simulate.H.singleparam}
\alias{simulate.H.singleparam}
\title{A function to simulate many random host phylogenies}
\usage{
\method{simulate}{H.singleparam}(tmax = 0, lambda, muH, K, timestep, reps,
  filename = NA)
}
\arguments{
\item{tmax:}{maximum time for which to simulate}

\item{lambda:}{host speciation rate}

\item{K:}{carrying capacity for host species}

\item{muH:}{host extinction rate}

\item{timestep:}{timestep for simulations}

\item{reps:}{number of times to simulate this set of parameters}

\item{filename:}{name underwhich set of simulations and statistics will be saved}
}
\description{
A function to run a certain number of replicate simulations, save all the trees and output all stats.
}
\examples{
simulate.H.singleparam()
}
\keyword{Host}
\keyword{multiple}
\keyword{phylogenies}
