% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cophys_HP}
\alias{simulate_cophys_HP}
\title{A function to simulate many random cophylogenies and calculate statistics}
\usage{
simulate_cophys_HP(tmax = 0, lambda, mu, beta, gamma, sigma, nu, kappa, delta,
  K, timestep, reps, filename = NA)
}
\arguments{
\item{tmax:}{maximum time for which to simulate}

\item{lambda:}{host speciation rate}

\item{K:}{carrying capacity for host species}

\item{mu:}{host extinction rate}

\item{beta:}{parasite host jump rate}

\item{gamma:}{dependency on genetic distance for host jumps}

\item{sigma:}{probability of successful co-infection following host jump}

\item{nu:}{parasite extinction rate}

\item{timestep:}{timestep for simulations}

\item{reps:}{number of times to simulate this set of parameters}

\item{filename:}{name underwhich set of simulations and statistics will be saved}
}
\description{
A function to run a certain number of replicate simulations, save all the trees and output all stats.
}
\examples{
simulate_cophys_HP()
}
\keyword{Host-Parasite}
\keyword{phylogeny,}
\keyword{statistics}
