% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cophys_PonH}
\alias{simulate_cophys_PonH}
\title{A function to simulate many random coevolving parasite phylogenies on
pre-built host-trees and calculate statistics in parallel.}
\usage{
simulate_cophys_PonH(Htrees, fromHtree = NA, toHtree = NA,
  HtreesPhylo = NA, P.startT, beta, gamma, sigma, nu, kappa, delta, timestep,
  reps1 = 1, reps2 = 1, filename = NA, ncores = 1)
}
\arguments{
\item{Htrees}{pre-built host trees on which to simulate parasite trees}

\item{fromHtree}{starting host-tree}

\item{toHtree}{finishing host-tree}

\item{HtreesPhylo}{list of host trees converted to phylo format}

\item{P.startT}{the timepoint at which a parasite invades the host-tree}

\item{beta}{parasite host jump rate}

\item{gamma}{dependency on genetic distance for host jumps}

\item{sigma}{probability of successful co-infection following host jump}

\item{nu}{parasite extinction rate}

\item{kappa}{the probability that a parasite duplicates within a host species}

\item{delta}{the probability that a parasite is lost randomly in one of the
daughter host species during a cospeciation event}

\item{timestep}{timestep for simulations}

\item{reps1}{the number of starting points for the parasite trees}

\item{reps2}{the number of replicates per starting point}

\item{filename}{name underwhich set of simulations and statistics will be
saved}

\item{ncores}{the number of cores that will be used to run simulations in
parallel}
}
\description{
The following function simulates parasite phylogenetic trees on pre-built
host trees using parallel computing.
}
\examples{
Htree<-list(rphylo_H(tmax=10, K=10, export.format="Raw"),
 rphylo_H(tmax=10, K=10, export.format="Raw"))
simulate_cophys_PonH(Htrees=Htree, P.startT=5, beta=0.1, gamma=0.02, sigma=0,
 nu=0.5, kappa=0, delta=0, timestep=0.001,reps1=1,reps2=2, filename="file")
}
\keyword{Host-Parasite}
\keyword{multiple}
\keyword{parallel}
\keyword{phylogeny,}
\keyword{statistics,}
