% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cophys_PonH_Htrait}
\alias{simulate_cophys_PonH_Htrait}
\title{A parallel parasite tree building function with host response to infection}
\usage{
simulate_cophys_PonH_Htrait(tmax, Htrees, HtreesPhylo = NA, fromHtree = NA,
  toHtree = NA, beta = 0.1, gamma = 0.2, sigma = 0, nu = 0.5,
  epsilon.1to0, epsilon.0to1, omega, rho, psi, TraitTracking = NA,
  prune.extinct = FALSE, export.format = "Phylo", P.startT = 0,
  reps1 = 1, reps2 = 1, ini.Hbranch = NA, Gdist = NA,
  timestep = 0.001, filename = NA, ncores = 1, DBINC = 100)
}
\arguments{
\item{tmax:}{maximum time for which to simulate}

\item{Htrees:}{a pre-built host phylogenetic tree}

\item{beta:}{parasite host jump rate}

\item{gamma:}{dependency on genetic distance for host jumps}

\item{sigma:}{probability of successful co-infection following host jump}

\item{nu:}{parasite extinction rate}

\item{epsilon.0to1:}{the baseline rate that a host with trait value 0 will mutate to a host with trait value 1}

\item{epsilon.1to0:}{the baseline rate that a host with trait value 1 will mutate to a host with trait value 0}

\item{omega:}{factor by which switching between trait values is altered depending on the trait value of the host and presence of parasites}

\item{rho:}{factor by which parasite extinction rate increases in response to host resistance}

\item{psi:}{factor by which parasite host-jump success decreases due to resistance of the new host}

\item{prune.extinct:}{whether to remove all extinct branches defaulting to FALSE}

\item{export.format:}{either "Phylo" (exported in Ape Phylo format, the default setting)) or "Raw" (just a list of branches as used within the function itself)}

\item{reps1:}{the number of starting points for the parasite trees}

\item{reps2:}{the number of replicates per starting point}

\item{P.startT:}{the timepoint at which a parasite invades the host-tree}

\item{ini.Hbranch:}{the host branch from which the parasite invasion is initiated (defaults to NA)}

\item{Gdist:}{can input a pre-calculated distance matrix of the living host branches at time of infection (defaults to NA)}

\item{timestep:}{timestep for simulations}
}
\description{
The following function simulates a parasite phylogenetic tree on a pre-built host phylogeny.
}
\examples{
simulate_cophys_PonH_Htrait()
}
\keyword{Host-Parasite}
\keyword{phylogeny}
